OMAP3: wlan: Fix memory access violation in wlan driver
In cases where drvMain_InvokeAction() is delayed, the driver can
attempt to read pNewAction from the queue after it has already been
freed by drvMain_InsertAction() drvMain_InvokeAction(). This patch
implements a check of the signal status to catch that. It is
largely based on a patch provided by PMC (author unknown).

Change-Id: I2431ff22d8bdecd49e81a460a76ec9801d2800a7
Signed-off-by: Thad Phetteplace <thad.phetteplace@ti.com>