WLAN - wl1283: Avoiding kernel crash with os_SignalObjectWait timeout
In case os_SignalObjectWait does timeout there are possibilities of crash.

After os_SignalObjectWait timeout driver frees the command memory while the
command is still on the queue. So next time when the command is about to be executed
que_Dequeue will try to access the memory which is already freed and causes a crash.

This change is for avoiding this crash

Change-Id: Id4813a98beb4cb88f8a0cf359f067673015bb9c4
Author: Brian Chen <bchen@ti.com>
Signed-off-by: Vishal Mahaveer <vishalm@ti.com>