wl12xx: add version and timestamp strings (DEBUG)
This is part of R5_00_08 package.

In order to allow better version control, add version
and timestamp strings into the wl12xx module.

These strings will be printed on driver load, and when
reading the driver_state.

(this is not going to work well in compat, because it
copies all the files instead of compiling them in the
original tree, so the version of the compat tree will
appear as the driver version. we can workaround it
with the additional makefile hacks...)

Change-Id: I289e0c543dd9611114cd2cae1cf777cc3bdb6310
Signed-off-by: Eliad Peller <eliad@wizery.com>
Signed-off-by: Panduranga Mallireddy <x0125717@ti.com>