wlan - hostapd.conf: update interface name for jb
Use wlan0 interface name for softAP.

Previously using same interface name for station and softAP confused the
framework. This is no longer the case.

Another reason for using this interface name is due to the fact that we are
using PDK builds and we cannot change interface name in framework.

Change-Id: Ifde92ebbd0fc8ec21e45be71fd55911bc1c32e4e
Signed-off-by: Vishal Mahaveer <a0271468@ti.com>